backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
scipy
pdfminer.six
python-multipart
regex
sentence-transformers
torch
pymupdf